requirements.txt
Flask
Flask-Cors
ultralytics
tensorflow
mediapipe
numpy
opencv-python-headless
python-dotenv
PyMySQL
PyJWT